fix my coding standards
