--- src/main/java/Irwyn/Exceptions/IrwynException.java
@@ -1,15 +1,0 @@
-package Irwyn.Exceptions;
-
-/**
- * This class represents an exception for Irwyn chatbot.
- */
-public class IrwynException extends Exception {
-    /**
-     * Constructor for the IrwynException class.
-     *
-     * @param msg The message that details the exception.
-     */
-    public IrwynException(String msg) {
-        super(msg);
-    }
-}
