--- src/main/java/Irwyn/Exceptions/TaskException.java
@@ -1,16 +1,0 @@
-package Irwyn.Exceptions;
-
-/**
- * This class represents a task exception for Irwyn chatbot.
- */
-public class TaskException extends InputException {
-    /**
-     * Constructor for the TaskException class.
-     */
-    public TaskException() {
-        super(
-                "Invalid task type\n"
-                + "Please use either todo | event | deadline\n"
-        );
-    }
-}
