--- src/test/java/Irwyn/Commands/DeadlineCommandTest.java
@@ -1,27 +1,0 @@
-package Irwyn.Commands;
-
-import Commands.DeadlineCommand;
-import Misc.StorageManager;
-import Misc.Ui;
-import Irwyn.Tasks.TaskList;
-
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class DeadlineCommandTest {
-    protected StorageManager s;
-    protected TaskList t;
-    protected Ui ui = new Ui();
-
-    @Test
-    public void testAddDeadline() throws Exception {
-        String filePath = System.getProperty("user.dir") + "/storage/testTaskData.txt";
-        s = new StorageManager(filePath);
-        t = new TaskList(s.load());
-        DeadlineCommand deadline = new DeadlineCommand("deadline test /by 2024-02-04 18:00:00");
-        deadline.execute(t, ui, s);
-        String expectedOutput = "[D][ ] test (by: Feb 04 2024 18:00)";
-        String actualOutput = t.getTasks().get(0).toString();
-        assertEquals(expectedOutput, actualOutput);
-    }
-}
