+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Irwyn
+Hello! I'm irwyn
 What can I do for you?
 ____________________________________________________________
